[NFC][VPlan] Update test checks to allow non-numerical symbols in VPValue names
We should add names to the VPValues for better readability of the VPlan dumps
and to ease tests creation by generating meaningful names in the VPlan itself
and capturing them by something like update_test_checks.py utility.

Adding that functionality in a proper way requires some mid-size (at least)
changes so I don't plan to commit it soon enough. However, I'm going to keep the
patch floating in the review stage to be applied in local workspaces and for it
to not grow too big the tests' update can be submitted already.

Change-Id: I09d4061036e497ec60e46ddd7f0c730913053f2f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/203841
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>