SYCL] Fix assertion for sycl keyword using without -fsycl option.
This is to fix assertion when clang compile C++ head file without -fsycl

The problem is when __constant (one of sycl keyworld) used without
-fsycl
option, the __contant should not be keyword and keyword status should be
KS_Disable.  But it is wrongly set to KS_Future.

The root problem is the KEYSYCL is add to KEYALLCXX:

KEYALLCXX = KEYSYCL | KEYCXX | KEYCXX11 | KEYCXX2A

It seems, we want KEYSYCL to be part of feature of C++.  But which C++
should belong: C++11 or C++2a?  (for example KEYCX2A... vs concept)

For now we should just use -fsycl option to control this.

This has been merged in into sycl
https://github.com/intel/llvm/pull/445

Change-Id: I5d6175f6af9cb41b4789b8a5ff6258879159fc77
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/204921
Tested-by: Yu, Jennifer <jennifer.yu@intel.com>
Reviewed-by: Keane, Erich <erich.keane@intel.com>
Reviewed-by: Prokofiev, Pavel G <pavel.g.prokofiev@intel.com>