Fix build fails of channels using -g with FPGA simulator on Windows
This issue is related to [CORC-4692]. The -g option triggers the
run of DebugInfoPass in the backend. Actually this error also
happens on Linux when setting CL_CONFIG_DBG_ENABLE=1.

The DebugInfoPass inserts debug info to all non-empty functions
in IR. This is implemented by inserting instructions like
"call get_*_id" and following with debug calls. However, the
follwing Barrier pass will add patch to syncless functions.
The DebugInfo pass inserts callInst into the pipe_ctor function
(the constructor of pipes, this is generated by the backend),
but fails to insert the following debug calls, resulting in that
the Barrier pass takes the pipe_ctor function as a syncless
function and patch it (this gives a function undefined error).

The patch modified DebugInfoPass so that the two operations, i.e.,
1. insert the callInst (insertComputeGlobalIds)
2. insert the following debug calls (addDebugCallsToFunction)
are processed on a function simultaneously.

Change-Id: I4d9ea758886ecc5abba40d62dad809dfc3f63ab5
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/205251
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>
Reviewed-by: Savonichev, Andrew <andrew.savonichev@intel.com>
Reviewed-by: Tianyang Zhu <tianyang.zhu@intel.com>
Tested-by: Xiebing Wang <xiebing.wang@intel.com>