[VPlan] Changes to fixup uniform inner loop exit condition check for loops under a predicate.
For uniform inner loops under a predicate, we always go through the inner loop
iterations even if the surrounding predicate is all-false due to our current
linearization. One way to fix this is using an allzero bypass around such loops.
The fix here fixes up the loop backedge check using the block predicate of the
inner loop preheader. Essentially, if the preheader predicate is all zeros we
should bail out of the inner loop immediately. Note that we would still iterate over
the inner loop once, but the code should be functionally correct since the loop body
would be masked properly.

Change-Id: Ib6bf926d20c81b63999e3d97b45371268e5bf46f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/205390
Tested-by: Guggilla, Satish <satish.guggilla@intel.com>
Reviewed-by: Kurkov, Vasiliy A <vasiliy.a.kurkov@intel.com>