[NFC][VPlan] Add a name member to the VPValue
Temporary solution to make dumps more readable and to make the tests easier to
write. Proper fix is to store the name in some map inside context similar to
llvm::Value way to do that. That would also allow to ensure uniqueness of the
names in a better way, e.g. consecutive numbering instead of re-using lower
bits of the address.

Change-Id: Ibbbae28dfc39f0d754e83dba31da9df607c6e958
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/203840
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>