[VPlan][Predicator] Fix block predicate of an early exit block for search loops
Before this change we were emitting wrong predicate for the early exit block in
the search loop, which was probably caused by the fact that the original version
of the code in predicator didn't expect to work on non-canonicalized loops with
multiple exits (all search-loop vectorization support is hacky still).

Fixing just the predicator would have caused the failure in that hacky CG
support for search loops vectorization and some change in the predicator is
needed to start preserving uniform control flow. As such, this commit tries to
isolate small independent piece of a change to ease review.

Change-Id: I43385dd4bf4e026c1608f8008d60700957aac873
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/206948
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>