[VPlan][Predicator] Emit not/and instruction closer to the block having the condition bit
That will be needed to properly maintain SSA form after starting to preserve
uniform control flow.

Not instruction is emitted in the block that has the condition (or plan entry
for external values). And is emitted in a newly introduced block right after the
one containing condition bit, so that the "and" instruction isn't predicated but
still located closer to the possible definition point.

Change-Id: Ic9c7f22da65b8f9ff98dffb78dd1b98956071b96
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/206952
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>