Add alternate ILPL codegen without completion buffers
This is a revival of Pablo's "Implemented ILPL without completion
buffer" patch. This was under review in the split repo as:

- commit b35e8ae1de5cc43b473f7fbe5d7ec5d74d323ff5
- change ID If695b4808e3db57beb6d5bb43cf03fcc463f1a64

The current ILPL code generation allows multiple outer loop instances to
use the units implementing the inner loop by reordering the outputs of
the inner loop with completion buffers. The completion buffer tokens
also limit the concurrency in the inner loop.

By contrast, the new approach added with this patch avoids the need to
reorder inner loop outputs by instead "spinning" any relevant outer loop
context through the inner loop. A deep LIC with initialized values
acting as a semaphore eliminates the need to use completion buffers at
all in this approach.

The goal of b35e8ae1 was to enable the new codegen and automatic ILPL in
general. This patch is less ambitious: ILPL remains manually-enabled by
default, and when it is enabled it still uses the completion buffer
approach. The only effect of this change should be that the code
generation can be switched to the new approach if the
"-csa-ilpl-use-completion-buf=false" option is given. A primary reason
for this is that the new spinning implementation is known to have bugs.
A secondary reason is that the performance impact is not well studied
yet.

Change-Id: I9399cd0577cdf0d0d76bab7a28ce581611fe2931
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/206969
Reviewed-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>
Tested-by: Creech, Timothy M <timothy.m.creech@intel.com>