CMPLRLLVM-10182: Disable optimizations wrapped in wholeprogramsafe intrinsic at -O0 [XMAIN]
The whole program devirtualization can run at O0. Intel's version of devirtualization
wraps the type metadata with the wholeprogramsafe intrinsic at compile time. If
the devirtualization flag is turned on, then devirtualizaion can be applied at O0
because the metadata is presented in the IR.

This change set will read the optimization level and lower the intrinsic
llvm.intel.wholeprogramsave into false. The constant propagation will clean the
IR and remove the basic block containing the metadata for devirtualization.

Change-Id: Ia902d32e84364c03c52c2d268859b94702a5637f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/207023
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>