[VPlan][VLS] Remove last remnants of DA from VLS
Initially, VPlanVLSAnalysis was expected to be based on DA. However, it
turned out that DA is not enough to provide all the necessary information
to the VLS analysis (e.g. the distance between two different pointers).
That's why implementation of VLS switched to using SCEV.

The patch enables better grouping for a few test cases, where DA failed
to recognize strided accesses. Now as we switched to SCEV, we can prove
that those accesses are strided and group them.

Change-Id: Ia746b3bf1793b18b3ae02a3878fecd4e32878e7c
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/209066
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>
Tested-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>