CMPLRLLVM-10210: Respect canExitEarly in case the user requests it [LLD]
Each driver can be called from another function if the developer wants it.
For example:

  foo() {
    lld::coff::link(args, canExitEarly, diag);
  }

If the user needs to collect data from LLD's global variables then it
will never happen because the branch in canExitEarly will never be taken.

This change set enables the canExitEarly, which produces the same path
as the community version of LLD. This patch won't be a problem for the
memory issue (CMPLRLLVM-10208) in llvm-lit. The testing system will
always set canExitEarly as 0, which mean the destructors of the global
variables will be invoked.

Also, this change includes the option intel-embedded-linker, which
prevents invoking llvm_shutdown. This is to prevent destroying any
data related to LLVM compiler and it can be user later in case
LLD is used embedded in another code.

Change-Id: If7824bcd374590eb6697e5a674e40d1df5bd9137
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/207471
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>