[FPGA] Fix blocking pipe built-ins import
OpenCL RTL doesn't contain implementation of blocking pipe built-ins,
so during built-ins import we are importing non-blocking pipes instead
of blocking and then renaming the imported functions into blocking, like:
  __read_pipe_2_bl_AS1  - function, that presents in the destination module;
  __read_pipe_2_fpga    - function, that presents in RTL module and going
                          to be imported;
  __read_pipe_2_bl_fpga - renamed built-in function, that is going to be
                          used in place of __read_pipe_2_fpga;

When a module contained both blocking and non-blocking types of the
built-ins, LLVMEqualizer was replacing non-blocking calls and declarations
with blocking in the destination module. That was happening, because the
algorithm was searching for all the functions in the destination module
having the same names as in RTL module and renames them.

This was fixed. We no longer do declaration import with following renaming
for blocking pipes - we just create a new one.

Change-Id: I4591da119f67b9a7f788f4c964c2216cdf1d6417
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/207570
Tested-by: Sidorov, Dmitry <dmitry.sidorov@intel.com>
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>