CMPLRLLVM-10353 - Windows compilations do not work form network drives
This is a cherry pick of the revert that caused the issue.  This is only going
into xmain-rel as we should acquire this fix naturally with pulldowns

---- original text from revert follows ----

Revert r361885 "[Driver] Fix -working-directory issues"

This made clang unable to open files using relative paths on network shares on
Windows (PR43204). On the bug it was pointed out that createPhysicalFileSystem()
is not terribly mature, and using it is risky. Reverting for now until there's
a clear way forward.

> Currently the `-working-directory` option does not actually impact the working
> directory for all of the clang driver, it only impacts how files are looked up
> to make sure they exist.  This means that that clang passes the wrong paths
> to -fdebug-compilation-dir and -coverage-notes-file.
>
> This patch fixes that by changing all the places in the driver where we convert
> to absolute paths to use the VFS, and then calling setCurrentWorkingDirectory on
> the VFS.  This also changes the default VFS for `Driver` to use a virtualized
> working directory, instead of changing the process's working directory.
>
> Differential Revision: https://reviews.llvm.org/D62271

This also revertes the part of r369938 which checked that -working-directory works.

Change-Id: Id261535b7ff94e09ec8605b5c200032ed7d3ea85
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/210363
Reviewed-by: Toguchi, Michael D <michael.d.toguchi@intel.com>
Tested-by: Toguchi, Michael D <michael.d.toguchi@intel.com>