[SYCL] Improve efficiency of stream output of h_item for FPGA.
This is done by creating a loop in writeHItem and ItemToStr,
rather than calling the same function multiple times.
Because everything is inlined in the FPGA flow, the latter results in many
copies of each function in hardware, which costs significant
FPGA area.  In a loop, the function will be created once on the
FPGA, and invoked multiple times.

Cherry-pick from intel/llvm

Change-Id: Iad8f4897118f70ee19c11c53478d64c4086c5232
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/209736
Reviewed-by: Bader, Alexey <alexey.bader@intel.com>
Tested-by: Maksimova, Viktoria <viktoria.maksimova@intel.com>
Reviewed-by: Kurkov, Vasiliy A <vasiliy.a.kurkov@intel.com>