[VPlan][Predicator] Fix bug in predicate calculation
When determining if condition in PredicateTerm should be negated we compared
direct successor with CurrBlock (which was true when PredicateTerm was formed
based on predecessors). However, once we started to use PostDomTree the check
should use VPPostDomTree instead. Fix this issue now (was lost during original
patch series rebase due to lack of a good testcase).

Change-Id: Ib583e31bd0a255a3ff6de3c8e07ca6ccb6ba2cdb
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/208856
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>