[CMPLRLLVM-10228] Update PGO metadata values when redundant cases are removed from switch instructions
A recent change is the IR verifier has started checking that the
branch_weights metadata for switch instructions has the correct
number of values for the number of branches in the switch instruction.
The INTEL_CUSTOMIZATION of updating switch instructions to eliminate cases
that are redundant with the default case (CQ368652) was not updating
the branch weight metadata. This change handles the updating of the
metadata to keep it consistent after cases are removed.

Change-Id: I700e09d98e3e0f327c78fae18332521096e438a0
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/209072
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>