[CMPLRLLVM-10325/CMPLRLLVM-10333] DTrans analysis needs to consider call made using Alias value
The stripPointerCasts() family of functions has been changed in the
community version to no longer look through aliases. DTrans analysis
was relying on this functionality to identify called functions for
safety analysis. Without this functionality, some structures that
were previously treated as safe will be marked as violating some
safety conditions.

This change modifies the locations in the analysis to try to
identify the called function by looking through a GlobalAlias,
rather than relying on stripPointerCasts() to do it.

Also, fixes a bug where bitcast alias would not be properly
analyzed for possible mismatched argument types.

Change-Id: I58a072ba1c6dfbd9c7a582dcc40a1a5d8986713b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/209662
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>