[NFC] Remove obsolete HIR VLS client
The old HIRVLSClient has never been fully implemented, there is almost
no tests for it and no plans to use it in the future. The patch
removes HIRVLSClient, so that we don't need to maintain it anymore
every time we make changes to VLS infrastructure.

Now the vectorizer uses VPlanHIRVLSClient.

Change-Id: I45b6dc1ed464f4a326fd5a93afd59c656af72867
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/210538
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>
Tested-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>