[intel_subgroups] Implement intel_sub_group_ballot
Implement ballot in 2 variants:
the first one is NEO-like, that is
extern "C" uint intel_sub_group_ballot(bool valid);
which is needed until NEO implements subgroups_next spec.
It is also mapped in Volcano.

The second implements subgroups_next variant that is planned to be a
Khronos extension:
__attribute((overloadable)) uint4 intel_sub_group_ballot(int predicate);
This variant is not supported in Volcano.

Change-Id: If38cc1ec94ad64f4a4c01ece49dde131b5dd29cb
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/208375
Tested-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Reviewed-by: Sachkov, Alexey <alexey.sachkov@intel.com>