[VPlan] Add all-zero bypass support.
This patch adds all-zero bypasses for non-uniform regions consisting of one or
more VPBasicBlocks. Regions are formed using the same block-predicate or where
the block predicate is anded with other predicates. Bypasses are also added for
non-uniform loop regions. TODO: proper cost modeling. As of now, a simple switch
is used to determine region size.

Change-Id: Ia0f6ba0444aacf2398ce9634f30de7c7d97d29ad
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/210359
Reviewed-by: Masten, Matt <matt.masten@intel.com>
Tested-by: Masten, Matt <matt.masten@intel.com>
Reviewed-by: Su, Ernesto <ernesto.su@intel.com>