[VPlan] Support unit stride load/store in LLVM-IR VPValue-based codegen.
This patch extends VPlan's VPValue-based vector codegen to recognize
and emit wide loads/stores instead of gathers/scatters for unit-strided
memory accesses. DA's VectorShape knowledge is used to identify unit-
strided pointers, and we generate a scalar GEP during codegen for such
accesses. The optimization introduced in this patch is protected with
a command-line option that allows us to toggle between using this code
or VLS to recognize and optimize unit-stride loads/stores.

Additionally a minor bug fix in DA to identify reverse unit-stride pointers
is also implemented.

Change-Id: I8578e221ab437060d5859da5eb9d7f592ef07dd2
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/210854
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>