[VPlan] Fix scalar value extraction for uniform call with vector return type
This happens to be a follow-up for
  [VPlan] Use lane 0 when searching for uniform instructions cache in serialized context
  (https://git-amr-2.devtools.intel.com/gerrit/206570)

Our "isUniform" check was inconsistent between serializeInstruction and
getScalarValue. That is correctly handled in the VPValue-based CG as we simply
use DA without any tricks but not in the case with LLVM-based CG that doesn't
have a single unified interface.

Instead of trying to match them, which is error-prone (as this issue shows) and
requires some unclear handling for privates, just fix the general/pessimistic
case of getting the scalar value from the vector one (as Karthik suggested in
the previous fix, thanks!).

Change-Id: I6488dbff408313af7b9a47e892be9b556a235988
Jira: CORC-6010
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/210990
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>