Add a new preprocessor define for all XISA features indicating the compiler was compiled with support for the feature.
Use this to control the includes in immintrin.h to allow the features to be used with target attribute without needing to specify the features on the command line.

With this we can remove the FIXME and do the normal _MSC_VER and INTRINSIC_PROMOTE checks that we use for other headers. Just all hidden behind the SUPPORTED define.

Change-Id: I79e8b358b051c0c1da08f4f6334e4e21bad83e63
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/211299
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Luo, Yuanke <yuanke.luo@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>