[VPlan] Changes to fix store to a uniform location in VPValue based code generation
When generating vector code for a store to a uniform location, we were generating
a wide store instead of a scatter causing some tests to fail. We can actually generate
better code for such a store that is not under a mask. This will be done later when
VPValue based code generation is enhanced with scalar/uniform analysis.

Change-Id: I008d14a4572e44246e15ba27892cbc3171091504
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/211586
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Pushkar Ratnalikar <pushkar.v.ratnalikar@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>