Temporarily fixed LIT failures at 7bdad084294 Reland "clang-misexpect: Profile Guided Validation of Performance Annotations in LLVM".
Note that the only thing it did is to hack the argument handling to
get -Wexpect in Clang working. The community patch effectively makes -Wexpect
always enabled in Clang. It doesn't fix the crash when multiple
CompilerInvocation with -pgo-warn-misexpect are made in the same process, but
I think we can live with it as long as we don't use -Wexpect. If we really
want to fix it then something more need to be done in llorg.

Change-Id: Id21ef11e3b844554ad7ef8da046ef5942491163e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/212176
Tested-by: Tianqing Wang <tianqing.wang@intel.com>
Reviewed-by: Tianqing Wang <tianqing.wang@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/212332
Reviewed-by: Bakhvalov, Denis <denis.bakhvalov@intel.com>
Tested-by: Bakhvalov, Denis <denis.bakhvalov@intel.com>