[VLS][1/3] Lexical order of memrefs: Add InsertPoint to OVLSGroup
The ordering of memory references is lost when a group is created. However, it
is important that we move loads only upward and move stores only downward. This
patch adds InsertPoint field to OVLSGroup to remember where the group-wide
load/store should be emitted. The patch doesn't affect functionality, since
InsertPoint is always set to be equal to the FirstMemref yet.

Change-Id: I820ce9aa493de4275d3ede0bb8dee6b508586ddc
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/213296
Reviewed-by: Gupta, Garima <garima.gupta@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>