[VLS][3/3] Lexical order of memrefs: Set correct InsertPoint for OVLSGroups
This patch modifies both interface of VLS server and its internal data
structures in order to account for lexical order between incoming memory
references. With this change, the server is able to form a group at the
lexically first/last memory reference location, and correctly check that
it is safe to move other memrefs to InsertPoint location.

This patch enables optimization of:
1) Non-monotonic accesses (e.g. ld[3*i+1], ld[3*i], ld[3*i+2])
2) Repeating accesses (e.g. ld[3*i], ld[3*i+1], ld[3*i+1], ld[3*i+2])
3) Intermixed load and store accesses (e.g. ld[2*i], st[2*i], ld[2*i+1], st[2*i+1])

Change-Id: I476bfaa0337a3ac4c803fe127b7d04541cdba433
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/213298
Reviewed-by: Gupta, Garima <garima.gupta@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>