[VPlan] Uplift codegen support for privates in LLVM-IR to be fully VPValue-based.
This patch updates VPlan's VPValue-based code generator for LLVM-IR
to use VPLoopEntity instructions in order recognize and correctly
support loop private variables. We generate a vector of pointers
for the widened private alloca, which is then used inside the vector
loop for vector context. This patch also removes all occurences of
Legality in VPValue-based codegen.

Change-Id: Ib3ff2d56033ad37b4601ba36f5ae657e6c7a84f7
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/213229
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>