[VPlan] Add support for analysis of privates in Divergence Analysis.
DA has to correctly mark private variables as 'divergent' so that
VP-value-based codegen can generate correct code for loop-private
variables. This is only done for the first run of the DA, using a
switch.

The changes in this patch DONOT correctly identify the stride of the
accesses to private variables. The correct code for this would require
a more extensive modification of how we infer shapes for GEP's and other
related instructions.

Change-Id: I60b38b89a2447073e10e748c9e7a4bde79baa2ca
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/213549
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: Pushkar Ratnalikar <pushkar.v.ratnalikar@intel.com>
Reviewed-by: Pushkar Ratnalikar <pushkar.v.ratnalikar@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>