[OCLVecClone][NFC] Prevent potential IR corruption while removing truncating sequences.
A recent optimization for get_local_id was introduced in OCLVecClone
when max work group size is known to be less than 2 Gig elements. As
part of this optimization we remove trivial truncating sequences from
the IR and replace with appropriate instructions. This replace and
remove operation was being done while iterating the use-def chain
which could corrupt the IR since iterators are invalidated after
instruction removal. This patch moves this operation to a separate
step after the truncating sequences are collected to prevent such
corruption.

Before this patch, the below LIT test was already failing on a RHEL8
machine (scel67, xmainoclefi2linux_debug) -
OCLBETests :: vectorizer/OCLVecClone/OptimizeGetLocalIDBuiltin.ll
Verified locally that the test passes after this fix.

Change-Id: I7c8b239f6b5caadbd0bddd8e3e6c08242f486b7e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/213631
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>