[VPlan] Read trip count metadata during HCFG creation
Instead of lazily computing the trip counts in VPLoopAnalysis[Base], read
everything during HCFG creation and store the data inside the latch
VPBasicBlock.

We don't want to store the data in VPLoop because ideally VPLoopInfo should be
able to re-analyze loop structure, but trip counts are property of the
underlying IR metadata, so should be read just once. As such, just try to mimic
what LLVM IR is doing.

Change-Id: If18de6a843c8713acae9ea435de4f1e70af1c86a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/213762
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>