cmplrllvm-9839:Fix access nullptr.
The natural of the problem is access nullptr.  The root problem is for
ThisExpr the (E)->getDecl() return NULL, since there no decl for ThisExpr.
This cause the call to cast<DeclRefExpr>(E)->getDecl() segfault.

To fix this just check ThisExpr, set nullptr.

Where passing nullptr for ThisExpr to decl_component_lists to get
components.

Note, for map clause:
Only allow expression containing only member access and/or array section
based on named variables.

And this pointer with array section only accept for [0:1], otherwise
error out.

With this restriction, for ThisExpr, only one array section can be accept,
so it is safe to pass nullptr for it at runtime point.

Change-Id: I6a40e28042aa23d92faae85a642e0d0d2ed073ad
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/214044
Reviewed-by: Rice, Michael P <michael.p.rice@intel.com>
Tested-by: Yu, Jennifer <jennifer.yu@intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>