CMPLRLLVM-10076: Compute argument alignment
The following change set targets two tickets, but one is related to the
other. The memory alignment for the first argument is being computed
in speq_qsort (505.mcf). This alignment is constant and we can collect
it from the allocation sites. For example, in spec_qsort we can see the
following:

   %2 = ptrtoint i8* %0 to i64
   %3 = and i64 %2, 7
   %4 = icmp eq i64 %3, 0
   br i1 %4, label %if_bb, label %end

This means that some modulo computation is happening against 8. The
argument %0 is defined as a calloc with size 8, or alloca with alignment
16. Which means that the ptrtoint instruction will produce a number that
is multiple of 8. The AND operation will produce 0 (%3) and the icmp will
always be true (%4).

This change set traces the module computation and verifies that %0 was
created with an alignment that is a multiple of 8. If so, then replace the
variable in the AND operation with a constant. The result will be the
following:

   %2 = ptrtoint i8* %0 to i64
   %3 = and i64 8, 7
   %4 = icmp eq i64 %3, 0
   br i1 %4, label %if_bb, label %end

The instruction simplification and callgraph simplification passes will be
called as part of the transformation to clean the IR. The result will be this:

   br label %if_bb

Change-Id: I3bfb4ac58914d7d94fe97419688113bdfc074a85
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/215450
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>