Enable ABI-compliant type coercion for FPGA emulator on Linux x64
This issue is related to [CORC-5785]. The target is to enable the
ABI-compliant type coercion (i.e. enable the CoerceTypes pass)
for FPGA emulator. The CoerceTypes pass is used to coerce the
aggregate types that should be passed in registers to those that
are guaranteed to be.

If the CoerceTypes pass is enabled, the error occurs when using
channel to transmit struct-type data on FPGA. In this case, a
@_z19_write_channel_intel_* function will be generated in the IR.
Once the type coercion is done, the number of arguments passed
to this @_z19_write_channel_intel_* function might change from
2 to 3. Since the following ChannelPipeTransformation pass only
handles the default case for channel write built-in functions
(i.e. the number of arguments is 2), the third argument of the
function will be ignored, which results in incorrect data or
channel read hang when trying to fetch data from the channels.

This patch merged the arguments of channel write functions when
arguments are split, i.e., the number of arguments after type
coercion changes from 2 to 3. The ChannelPipeTransformation pass
is modified so that the correct data pointer is passed to the
transformed pipe write function.

Change-Id: I9b3ed29f6f44a8fcae7b8575e4b7535ef6357f3f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/214357
Tested-by: Xiebing Wang <xiebing.wang@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>