[VPlan] Add support for masked reductions recognized by HIRSafeReduction analysis.
This patch extends VPlan's VPEntities framework to correctly analyze
and represent masked safe reductions as VPReductions. The analysis to
determine the starting PHI of a reduction is modified to be recursive
for PHI users of exiting instructions, along with updating the exit
instruction to appropriate blend PHI. The patch also enables CG
support for such masked reductions in mixed CG approach only. Support
for full VPValue-based CG is a TODO.

Change-Id: I45827c17b3f2f42ce2a32cd6cd7be07ab1d90964
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/214548
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>