Fix assertion failure of check-ocl-runtime on Win Debug mode
This issue is related to [CORC-6278]. The patch fixed the
assertion failure of ocl runtime check test. The error occurs only
in debug mode on Windows and is not reproduced on Linux.

In debug mode, the assertion statement takes efforts and the OpenCL
runtime unit test FrameworkTestType.Test_clLinkProgramTest will
gives assertion failure. This is caused by the mutex and the
Error_handler assertion failuire. (1) The mutex is used in
CompileService.cpp and locked before calling BuildProgram()
function. Originally this mutex is locked in try{} code block, in
this case, the mutex cannot be unlocked if an exception is caught.
(2) The Error_handler is used in ProgramBuilder.cpp and it is also
registered in try{} code block, which results in similar issue as
the mutex above.
If an exception is caught, the execution flow will jump out of the
try{} code block. In this case, the mutex is still locked and the
Error_handler remains registered, which causes the following
assertion failure when the mutex is destroyed and the Error_handler
is removed. This is observed on Windows with VS.

This patch fixes the assertion failure when handling these two
variables. For the mutex, the lock statement is moved out of the
try{} code block. For the Error_handler, it is de-registered in the
catch{} code block.

Change-Id: Ieb8be5cf9f3586b1cea5ba61d6cefac20431f58c
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/214672
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>