[VPlan] Corrected VPValue-based code gen for reduction phi.
An incorrect predicate function is called to detect whether a PHI node
is a reduction-PHI, i.e. a recurrent PHI node. The predicate called
returns true if the predicated value is one of the instruction linked with
a reduction.
Created the more exact predicate that makes additional check for that the node
passed is a recurrence-PHI

Change-Id: I9ee655ffc1c9f281c03f0b5a3e08af43e471891d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/215479
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Tested-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>