CMPLRLLVM-10672: Clean up for whole program trace (Part 1)
The current whole program trace depends on an mllvm variable.
This change set moves this trace to -debug-only=whole-program-analysis.
It also contains a cleanup in the trace itself. All the trace related
to whole program is encapsulated in one function.

The change set also fix an issue with whole-program-assert. It should be
calling LLVM error rather than assert itself.

This change separated some test cases that were testing for debug trace and
IR at the same time.

Change-Id: I35f4eee0d507ca279de954e8ed83c1155ec18445
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/215731
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>