[NFC][VPlan] Test changes for enabling VPValue-based code gen.
All test changes were related to the introduction of new VPInstructions for reductions
under -enable-vp-value-codegen. The old run line was not preserved because the intent
of the tests is still the same, and that is to check the HCFG for proper SSA and that
the inner loop control flow massaging is correct. This will be the same regardless of
whether or not the new VPInstructions are present.

Change-Id: Ibe1042dfc45f4e444fd84dd7b106e6fc79a26ef3
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/215756
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>