[NFC][VPlan] Changes to 10 LIT tests for VPValue CG enabling.
* 'vplan_svml_fn_lookup.ll' requires additional changes in VPOCodegen.cpp,
   (CMPLRLLVM-10673).
* 'vplan_vpo_vectorize_extract_subvector.ll': In VPValue codegen, we are
  generating 'call' instead of 'tail call'. Might require some more
  investigation.

I also take this opportunity to reduce size of some of the tests (More
than one was submitted by me) and clean-up a little. The cleanup is to
remove unnecessary metadata, attributes, unused function decls and
minimize the IR in the original test when it is unrelated to the issue
being addressed.

Change-Id: Iba8bf69e2c799270cf48fc88f3cb90635181cbba
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/215609
Reviewed-by: Pushkar Ratnalikar <pushkar.v.ratnalikar@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>