CMPLRLLVM-9096 - Implement preliminary support for the OpenMP tile clause
Adds a 'tile' clause to 'omp for' loops (including combined forms). The
clause specifies a list of non-negative integral expressions the number
of which determines the number of loops that apply to the OpenMP loop. Any
collapse clause specified is ignored (with warning). An implicit collapse
clause is added with the number of tile expressions to trigger the
processing of an uncollapsed loop.

Change-Id: Id2c2f356860b91df6d2db87bcdae4fd5b4c107a8
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/217386
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Yu, Jennifer <jennifer.yu@intel.com>
Reviewed-by: Narayanaswamy, Ravi <ravi.narayanaswamy@intel.com>